refactor(form): extract validators to separate module
Split form.rs (1749 lines) into modular structure:

- validators.rs: ValidationError, ValidatorFn, Validators
- mod.rs: FormField, FormFieldBuilder, FormState, FormStateBuilder

This improves code organization while maintaining the same API.

Closes #198 (partial)
